+++ command/example.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-options="-Wall -Wextra -pedantic -std=c99 -g -fsanitize=address -fsanitize=undefined"
+options="-I. -Wall -Wextra -pedantic -std=c99 -g -fsanitize=address -fsanitize=undefined"
 macos_suppress_options="-Wno-deprecated-declarations -Wno-c11-extensions"
 
 if [[ "$OSTYPE" == "darwin"* ]]; then